+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.15.1]
 
 ### Changed
-- Nothing yet
+- - Implemented Drop for AsyncStream (TSP-584)
 
 ## [0.15.0]